incubadora/requirements.txt
rich
groq
google-api-python-client
moviepy
rich
python-dotenv
requests
boto3
instagrapi
pandas
paramiko
